Puntos decimales añadido

+++ Actividad1/src/App.java
@@ -1,5 +1,4 @@
 import java.util.Scanner;
-
 import javax.swing.JOptionPane;
 
 public class App {
@@ -7,11 +6,11 @@ public class App {
         
         Scanner scanner = new Scanner(System.in);
         String x = JOptionPane.showInputDialog("Ingrese el primer número: ");
-        int num1 = Integer.parseInt(x);
+        float num1 = Float.parseFloat(x);
         String y = JOptionPane.showInputDialog("Ingrese el segundo número: ");
-        int num2 = Integer.parseInt(y);
+        float num2 = Float.parseFloat(y);
         String z = JOptionPane.showInputDialog("Ingrese el tercer número: ");
-        int num3 = Integer.parseInt(z);
+        float num3 = Float.parseFloat(z);
 
         float promedio = (num1 + num2 + num3) / 3;
         String nombre = JOptionPane.showInputDialog("Ingrese su nombre: ");
